RavenDB-14779 - make ArrayStream public
